Introduce the new variables based on percentage.
After some tests, we found that an appropriate bucket span would be 4 hours for a 7 days dashboard time range.
By making simple conversion factors, we come up with a 2.4% bucket span of the dashboard time range. Since the intention of the two variables is to get the span in which the event is more or less in the middle of the range, the __bucket_from variable is an 98.8% of the total time range, and the __bucket_to variable is a 101.2% of the total time range, building at the end a 2.4% bucket span.