Create percentual bucketspan vars (#242)
In some cases scenarios, the status panel is used to go from one dashboard to a more specific one, e.g. when performing root/cause analysis. In such cases, the user may want to reduce the scope of the time span, while keeping the desired event centered (to be able to analyse the previous/posterior buckets of time).
The status panel does not have dynamically arranged variables to do so, and this is what this PR tries to solve. It is a super-duper simple implementation which comes from the experience. After some tests, we found that 4h bucket span works perfectly with a 7 days dashboard time range, which is a 2.4% of it.
This way, the scope and time span of the dashboard can be reduced, acting like a "zoom".

Co-authored-by: alberto.perdomo <alberto.perdomo@datadope.io>